B1.1.1SM - Sim working and some math
When trying to simulate code, the AdvantageKit template was forcing me to try to read from a log file. I overwrote this in Robot.java by making the if statement true that checks if the robot is in Sim mode.

Also, I added a method to verify our cartesian-to-polar method was working correctly by converting from polar-to-cartesian. This test verified the method, but I am going to leave it in as a helpful tool.