B1.2.5AT - cleaned up folders a bit more, & more swerve.
I added the arm commands into their own folder so we're not looking at a million commands at once. Also added some important printouts for Swerve and a command that will put the wheels at a specific angle. We should start with this command and then move onto driving. The Field2D object is also added and we should be able to get data in SmartDashboard and a visualization in Glass.